ctrl+z'd change accidentally. JSON response takes an int64, not an int

+++ api/request.go
@@ -9,7 +9,7 @@ import (
 )
 
 type response struct {
-	Prime int `json:"primeNum"`
+	Prime int64 `json:"primeNum"`
 }
 
 func GetPrime(c *gin.Context) {
